+++ CHANGES.md
@@ -9,6 +9,7 @@ Changes in 5.4
 - Remove / deprecate $type param from ColorInput() in Inputs.php & program wide
 - CSS add input[type=color] & .color-input-value in stylesheet.css
 - Add Match password action hook in Password.php & Actions.php
+- Add Student Billing alerts in Portal.php & Student_Billing/includes/PortalAlerts.fnc.php
 
 Changes in 5.3
 --------------